[BE] Station, Payment 유효성 검증, ExceptionCode 정리,  임시 Cors 컨트롤 webConfig, + 단일 마커 조회용 API 추가 (#90)
* Docs: data.sql 수정
* Fix: ExceptionCode 명확화
* Fix: Payment, Station 유효성 검증 추가
* Fix: Payment 유효성 검증 적용
* Fix: 위/경도 double로 타입변경
* Fix: Station에 phone 필드 추가
* Docs: 세미콜론 추가
* Feat: Cors용 임시 webConfig
* Feat: 단일 Station 가용 배터리 조회 API (마커 클릭용)
* Docs: 단일 Station 가용 배터리 조회 API용 더미 Reservation 수정

+++ backend/src/main/java/backend/domain/station/entity/StationLocation.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter @Setter
 public class StationLocation {
 
-    private long latitude;
+    private double latitude;
 
-    private long longitude;
+    private double longitude;
 
 }
